+++ NEWS.md
@@ -3,6 +3,7 @@
 * [MODQM-260] (https://issues.folio.org/browse/MODQM-260) - Supports users interface 15.0 16.0
 * [MODQM-271] (https://issues.folio.org/browse/MODQM-271) - Extend GET /records endpoint to support bib-auth links
 * [MODQM-291] (https://issues.folio.org/browse/MODQM-291) - Upgrade to folio-spring-base v5.0.0
+* [MODQM-281] (https://issues.folio.org/browse/MODQM-281) - Remove 008 Desc from request/response on edit/derive MARC bib
 
 ## 2.4.0 - Released