+++ NEWS.md
@@ -5,6 +5,7 @@
 ## 2.5.0 - Released
 
 ### Changes
+* [MODQM-253](https://issues.folio.org/browse/MODQM-253) - Changes to Leader position 17 validation AND 008 byte
 * [MODQM-260](https://issues.folio.org/browse/MODQM-260) - Supports users interface 15.0 16.0
 * [MODQM-291](https://issues.folio.org/browse/MODQM-291) - Upgrade to folio-spring-base v5.0.0
 * [MODQM-281](https://issues.folio.org/browse/MODQM-281) - Remove 008 Desc from request/response on edit/derive MARC bib